Move common code to new helpers lua script. Resetting not null attributes for indoor and water
