+++ flex-config/all_tags.lua
@@ -1,3 +1,5 @@
+require "helpers"
+
 -- Put all OSM tag data into a single table w/out geometry
 local json = require('dkjson')
 
@@ -54,15 +56,6 @@ function all_tags_process_relation(object)
     end
 end
 
-
-
--- deep_copy based on copy2: https://gist.github.com/tylerneylon/81333721109155b2d244
-function deep_copy(obj)
-    if type(obj) ~= 'table' then return obj end
-    local res = setmetatable({}, getmetatable(obj))
-    for k, v in pairs(obj) do res[deep_copy(k)] = deep_copy(v) end
-    return res
-end
 
 
 if osm2pgsql.process_node == nil then
