+++ flex-config/infrastructure.lua
@@ -1,8 +1,9 @@
+require "helpers"
+
 -- Change SRID if desired
 local srid = 3857
 
 local tables = {}
-
 
 -- Rows with any of the following keys will be treated as possible infrastructure
 local infrastructure_keys = {
@@ -14,24 +15,7 @@ local infrastructure_keys = {
     'power'
 }
 
--- Function make_check_in_list_func from: https://github.com/openstreetmap/osm2pgsql/blob/master/flex-config/compatible.lua
-function make_check_in_list_func(list)
-    local h = {}
-    for _, k in ipairs(list) do
-        h[k] = true
-    end
-    return function(tags)
-        for k, _ in pairs(tags) do
-            if h[k] then
-                return true
-            end
-        end
-        return false
-    end
-end
-
 local is_infrastructure = make_check_in_list_func(infrastructure_keys)
-
 
 tables.infrastructure_point = osm2pgsql.define_table({
     name = 'infrastructure_point',
@@ -49,63 +33,6 @@ tables.infrastructure_point = osm2pgsql.define_table({
     }
 })
 
-
-
-function parse_height(input)
-    if not input then
-        return nil
-    end
-
-    local height = tonumber(input)
-
-    -- If height is just a number, it is in meters, just return it
-    if height then
-        return height
-    end
-
-    -- If there is an 'ft' at the end, convert to meters and return
-    if input:sub(-2) == 'ft' then
-        local num = tonumber(input:sub(1, -3))
-        if num then
-            return num * 0.3048
-        end
-    end
-
-    return nil
-end
-
-
--- Parse an ele value like "1800", "1955 m" or "8001 ft" and return a number in meters
-function parse_ele(input)
-    if not input then
-        return nil
-    end
-
-    local ele = tonumber(input)
-
-    -- If ele is just a number, it is in meters, so just return it
-    if ele then
-        return ele
-    end
-
-    -- If there is an 'm ' at the end, strip off and return
-    if input:sub(-1) == 'm' then
-        local num = tonumber(input:sub(1, -2))
-        if num then
-            return num
-        end
-    end
-
-    -- If there is an 'ft' at the end, strip off and return
-    if input:sub(-2) == 'ft' then
-        local num = tonumber(input:sub(1, -3))
-        if num then
-            return math.floor(num * 0.3048)
-        end
-    end
-
-    return nil
-end
 
 
 function infrastructure_process_node(object)
@@ -219,15 +146,6 @@ function infrastructure_process_node(object)
 
 end
 
-
-
--- deep_copy based on copy2: https://gist.github.com/tylerneylon/81333721109155b2d244
-function deep_copy(obj)
-    if type(obj) ~= 'table' then return obj end
-    local res = setmetatable({}, getmetatable(obj))
-    for k, v in pairs(obj) do res[deep_copy(k)] = deep_copy(v) end
-    return res
-end
 
 
 if osm2pgsql.process_node == nil then
