+++ flex-config/natural.lua
@@ -1,3 +1,4 @@
+require "helpers"
 
 -- Change SRID if desired
 local srid = 3857
@@ -42,40 +43,6 @@ tables.natural_polygon = osm2pgsql.define_table({
         { column = 'geom',     type = 'multipolygon' , projection = srid},
     }
 })
-
-
--- Parse an ele value like "1800", "1955 m" or "8001 ft" and return a number in meters
-function parse_ele(input)
-    if not input then
-        return nil
-    end
-
-    local ele = tonumber(input)
-
-    -- If ele is just a number, it is in meters, so just return it
-    if ele then
-        return ele
-    end
-
-    -- If there is an 'm ' at the end, strip off and return
-    if input:sub(-1) == 'm' then
-        local num = tonumber(input:sub(1, -2))
-        if num then
-            return num
-        end
-    end
-
-    -- If there is an 'ft' at the end, strip off and return
-    if input:sub(-2) == 'ft' then
-        local num = tonumber(input:sub(1, -3))
-        if num then
-            return math.floor(num * 0.3048)
-        end
-    end
-
-
-    return nil
-end
 
 
 function natural_process_node(object)
@@ -153,15 +120,6 @@ function natural_process_way(object)
         })
     end
     
-end
-
-
--- deep_copy based on copy2: https://gist.github.com/tylerneylon/81333721109155b2d244
-function deep_copy(obj)
-    if type(obj) ~= 'table' then return obj end
-    local res = setmetatable({}, getmetatable(obj))
-    for k, v in pairs(obj) do res[deep_copy(k)] = deep_copy(v) end
-    return res
 end
