+++ flex-config/place.lua
@@ -1,3 +1,4 @@
+require "helpers"
 
 -- Change SRID if desired
 local srid = 3857
@@ -107,15 +108,6 @@ function place_process_relation(object)
     end
 end
 
-
-
--- deep_copy based on copy2: https://gist.github.com/tylerneylon/81333721109155b2d244
-function deep_copy(obj)
-    if type(obj) ~= 'table' then return obj end
-    local res = setmetatable({}, getmetatable(obj))
-    for k, v in pairs(obj) do res[deep_copy(k)] = deep_copy(v) end
-    return res
-end
 
 
 if osm2pgsql.process_node == nil then
