+++ flex-config/traffic.lua
@@ -1,3 +1,5 @@
+require "helpers"
+
 -- Change SRID if desired
 local srid = 3857
 
@@ -227,15 +229,6 @@ function traffic_process_way(object)
     
 end
 
-
-
--- deep_copy based on copy2: https://gist.github.com/tylerneylon/81333721109155b2d244
-function deep_copy(obj)
-    if type(obj) ~= 'table' then return obj end
-    local res = setmetatable({}, getmetatable(obj))
-    for k, v in pairs(obj) do res[deep_copy(k)] = deep_copy(v) end
-    return res
-end
 
 
 if osm2pgsql.process_node == nil then
